pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reporesolve"
version = "0.1.0"
description = "Agentic supervisor for building and repairing multi-repository environments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "questionary>=2.0.1",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1",
]
authors = [
  { name = "Your Name" }
]

[project.scripts]
reporesolve = "reporesolve.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["reporesolve*"]